make two animation for summary plot
